requirements.txt
discord.py
ollama
Flask
flask-cors
Pillow
uvicorn
asgiref
python-dotenv
pre-commit
detect-secrets
